requirements.txt
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
